requirements-dev.txt
railroad-diagrams
cairosvg
pylint
rope